third_party/WebKit/LayoutTests/fast/storage/storage-detached-iframe-expected.txt

CONSOLE WARNING: Attr child nodes are deprecated and will be removed in M45, around August 2015. Please use 'Attr.value' instead.
CONSOLE MESSAGE: line 25: Expected exception caught.
CONSOLE MESSAGE: line 39: Expected exception caught.
Bug: https://bugs.webkit.org/show_bug.cgi?id=57140

Description: Crash from null pointer dereference below WebCore::StorageAreaImpl::setItem()

Expect: passed

Result: passed
